+++ src/draw-segment.js
